+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added [`search()`](./docs/jobs#search) function to Jobs class
+- Added `minEntries` and `maxEntries` to `File` element
 
 # 0.4.0 (2019-12-18)